refactor: move repo code into shared lib
