+++ modules/outbound-command-event-handler/src/application/handlers/process_party_info_callback.ts
@@ -25,38 +25,46 @@
 'use strict';
 
 import { ILogger } from '@mojaloop/logging-bc-public-types-lib';
-import { CommandEventMessage, ProcessSDKOutboundBulkPartyInfoRequestCompleteMessage } from '@mojaloop/sdk-scheme-adapter-private-shared-lib';
-import { BulkTransactionAgg } from '..';
-import { ICommandEventHandlerOptions } from '@module-types';
-import { BulkTransactionInternalState } from '../..';
+import { BulkTransactionAgg, CommandEventMessage, ProcessPartyInfoCallbackMessage, PartyInfoCallbackProcessedMessage, ICommandEventHandlerOptions, IndividualTransferInternalState } from '@mojaloop/sdk-scheme-adapter-private-shared-lib';
+import { v1_1 as FSPIOP } from '@mojaloop/api-snippets';
 
+type PartyResult = FSPIOP.Schemas.PartyResult;
 
-export async function handleProcessSDKOutboundBulkPartyInfoRequestCompleteMessage(
+export async function handleProcessPartyInfoCallbackMessage(
     message: CommandEventMessage,
     options: ICommandEventHandlerOptions,
     logger: ILogger,
 ): Promise<void> {
-    const processSDKOutboundBulkPartyInfoRequestCompleteMessage = 
-        message as ProcessSDKOutboundBulkPartyInfoRequestCompleteMessage;
+    const processPartyInfoCallbackMessage = message as ProcessPartyInfoCallbackMessage;
     try {
-        logger.info(`Got ProcessSDKOutboundBulkPartyInfoRequestCompleteMessage: bulkid=${processSDKOutboundBulkPartyInfoRequestCompleteMessage.getKey()}`);
+        logger.info(`Got ProcessPartyInfoCallbackMessage: id=${processPartyInfoCallbackMessage.getKey()}`);
 
         // Create aggregate
         const bulkTransactionAgg = await BulkTransactionAgg.CreateFromRepo(
-            processSDKOutboundBulkPartyInfoRequestCompleteMessage.getKey(),
+            processPartyInfoCallbackMessage.getBulkId(),
             options.bulkTransactionEntityRepo,
             logger,
         );
 
-        const bulkTx = bulkTransactionAgg.getBulkTransaction();
-
-        bulkTx.setTxState(BulkTransactionInternalState.DISCOVERY_COMPLETED);
-
-        if(bulkTx.isAutoAcceptPartyEnabled()) {
-            // TODO: construct and send SDKOutboundBulkAcceptPartyInfoRequested message to domain event handler
+        const individualTransfer = await bulkTransactionAgg.getIndividualTransferById(
+            processPartyInfoCallbackMessage.getTransferId(),
+        );
+        const partyResult = <PartyResult>processPartyInfoCallbackMessage.getContent();
+        if(partyResult.errorInformation) {
+            individualTransfer.setTransferState(IndividualTransferInternalState.DISCOVERY_FAILED);
+        } else {
+            individualTransfer.setTransferState(IndividualTransferInternalState.DISCOVERY_SUCCESS);
         }
+        individualTransfer.setPartyResponse(partyResult);
 
-        await bulkTransactionAgg.setTransaction(bulkTx);
+        const msg = new PartyInfoCallbackProcessedMessage({
+            key: processPartyInfoCallbackMessage.getKey(),
+            timestamp: Date.now(),
+            headers: [],
+        });
+        await options.domainProducer.sendDomainMessage(msg);
+
+        await bulkTransactionAgg.setIndividualTransferById(individualTransfer.id, individualTransfer);
     /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
     } catch (err: any) {
         logger.info(`Failed to create BulkTransactionAggregate. ${err.message}`);
