+++ modules/outbound-command-event-handler/src/application/handlers/process_sdk_outbound_bulk_party_info_request.ts
@@ -26,14 +26,14 @@
 
 import { ILogger } from '@mojaloop/logging-bc-public-types-lib';
 import {
+    BulkTransactionAgg,
     CommandEventMessage,
     ProcessSDKOutboundBulkPartyInfoRequestMessage,
     PartyInfoRequestedMessage,
+    ICommandEventHandlerOptions,
+    BulkTransactionInternalState,
+    IndividualTransferInternalState,
 } from '@mojaloop/sdk-scheme-adapter-private-shared-lib';
-import { BulkTransactionAgg } from '..';
-import { ICommandEventHandlerOptions } from '@module-types';
-import { BulkTransactionInternalState } from '../..';
-import { IndividualTransferInternalState } from '../..';
 
 export async function handleProcessSDKOutboundBulkPartyInfoRequestMessage(
     message: CommandEventMessage,
