+++ modules/outbound-command-event-handler/src/types/index.ts
@@ -22,6 +22,4 @@
  --------------
  ******/
 
-export * from './bulk_transaction_entity_repo';
-export * from './command_event_handler_options';
 export * from './repo_infra_types';
