+++ modules/outbound-command-event-handler/tsconfig.json
