+++ modules/private-shared-lib/package.json
@@ -32,6 +32,7 @@
     "@mojaloop/platform-shared-lib-messaging-types-lib": "^0.0.3",
     "@mojaloop/platform-shared-lib-nodejs-kafka-client-lib": "^0.0.8",
     "ajv": "^8.11.0",
+    "redis": "^4.3.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
