+++ modules/private-shared-lib/src/domain/bulk_transaction_entity.ts
@@ -24,7 +24,8 @@
 
 'use strict';
 
-import { BaseEntityState, BaseEntity, SchemaValidationError } from '@mojaloop/sdk-scheme-adapter-private-shared-lib';
+import { BaseEntityState, BaseEntity } from './';
+import { SchemaValidationError } from '../errors';
 import { SDKSchemeAdapter } from '@mojaloop/api-snippets';
 import { randomUUID } from 'crypto';
 import Ajv from 'ajv';
