+++ modules/private-shared-lib/src/domain/index.ts
@@ -43,3 +43,6 @@ export * from './base_aggregate';
 export * from './base_entity';
 export * from './base_entity_state';
 export * from './ientity_state_repository';
+export * from './bulk_transaction_agg';
+export * from './bulk_transaction_entity';
+export * from './individual_transfer_entity';
