+++ modules/private-shared-lib/src/infra/inmemory_bulk_transaction_repo.ts
@@ -24,19 +24,13 @@
 
 'use strict';
 
-import * as redis from 'redis';
 import { ILogger } from '@mojaloop/logging-bc-public-types-lib';
-import { BulkTransactionState, IndividualTransferState } from '../domain';
-import { IBulkTransactionEntityRepo } from '../types/bulk_transaction_entity_repo';
+import { BulkTransactionState, IndividualTransferState  } from '../domain';
+import { IBulkTransactionEntityRepo } from '../types';
 
-export interface IRedisBulkTransactionStateRepoOptions {
-    connStr: string;
-}
-
-export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo {
-    protected _redisClient!: redis.RedisClientType;
-
-    private readonly _redisConnStr: string;
+export class InMemoryBulkTransactionStateRepo implements IBulkTransactionEntityRepo {
+    /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+    private _data: any;
 
     private readonly _logger: ILogger;
 
@@ -44,25 +38,20 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
 
     private readonly keyPrefix: string = 'outboundBulkTransaction_';
 
-    constructor(options: IRedisBulkTransactionStateRepoOptions, logger: ILogger) {
-        this._redisConnStr = options.connStr;
+    private readonly individualTransferKeyPrefix: string = 'individualItem_';
+
+
+    constructor(logger: ILogger) {
         this._logger = logger;
     }
 
     async init(): Promise<void> {
-        this._redisClient = redis.createClient({ url: this._redisConnStr });
-        this._redisClient.on('error', err => {
-            this._logger.error(err, 'Error connecting to redis server: ' + err.message);
-            if(!this._initialized) {
-                throw (err);
-            }
-        });
-        await this._redisClient.connect();
+        this._data = {};
         this._initialized = true;
     }
 
     async destroy(): Promise<void> {
-        if(this._initialized) { this._redisClient.quit(); }
+        if(this._initialized) { this._data = undefined; }
         return Promise.resolve();
     }
 
@@ -76,15 +65,16 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
         }
         const key: string = this.keyWithPrefix(id);
         try {
-            const bulkTransactionEntityStateStr = await this._redisClient.hGet(key, 'bulkTransactionEntityState');
+
+            const bulkTransactionEntityStateStr = this._data[key].bulkTransactionEntityState;
             if(bulkTransactionEntityStateStr) {
                 return JSON.parse(bulkTransactionEntityStateStr);
             } else {
-                this._logger.error('Error loading entity state from redis - for key: ' + key);
-                throw (new Error('Error loading entity state from redis'));
+                this._logger.error('Error loading entity state from memory - for key: ' + key);
+                throw (new Error('Error loading entity state from memory'));
             }
         } catch (err) {
-            this._logger.error(err, 'Error loading entity state from redis - for key: ' + key);
+            this._logger.error(err, 'Error loading entity state from memory - for key: ' + key);
             throw (err);
         }
     }
@@ -95,9 +85,9 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
         }
         const key: string = this.keyWithPrefix(id);
         try {
-            await this._redisClient.del(key);
+            delete this._data[key];
         } catch (err) {
-            this._logger.error(err, 'Error removing entity state from redis - for key: ' + key);
+            this._logger.error(err, 'Error removing entity state from memory - for key: ' + key);
             throw (err);
         }
     }
@@ -108,47 +98,41 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
         }
         const key: string = this.keyWithPrefix(entityState.id);
         try {
-            await this._redisClient
-                .multi()
-                .hSet(key, 'id', entityState.id || '')
-                .hSet(key, 'bulkTransactionEntityState', JSON.stringify(entityState))
-                .exec();
+            this._data[key] = {
+                id: entityState.id || '',
+                bulkTransactionEntityState: JSON.stringify(entityState),
+            };
         } catch (err) {
-            this._logger.error(err, 'Error storing entity state to redis - for key: ' + key);
+            this._logger.error(err, 'Error storing entity state to memory - for key: ' + key);
             throw (err);
         }
     }
-    
+
+
     async getAllIndividualTransferIds(bulkId: string): Promise<string[]> {
         if(!this.canCall()) {
             throw (new Error('Repository not ready'));
         }
         const key: string = this.keyWithPrefix(bulkId);
         try {
-            const allAttributes = await this._redisClient.hKeys(key);
-            const allIndividualTransferIds = allAttributes.filter(attr => attr.startsWith('individualItem_')).map(attr => attr.replace('individualItem_', ''));
+            const allAttributes = Object.keys(this._data[key]);
+            const allIndividualTransferIds = allAttributes.filter(attr => attr.startsWith(this.individualTransferKeyPrefix)).map(attr => attr.replace(this.individualTransferKeyPrefix, ''));
             return allIndividualTransferIds;
         } catch (err) {
-            this._logger.error(err, 'Error getting individual transfers from redis - for key: ' + key);
+            this._logger.error(err, 'Error getting individual transfers from memory - for key: ' + key);
             throw (err);
         }
     }
-    
+
     async getIndividualTransfer(bulkId: string, individualTranferId: string): Promise<IndividualTransferState> {
         if(!this.canCall()) {
             throw (new Error('Repository not ready'));
         }
         const key: string = this.keyWithPrefix(bulkId);
         try {
-            const individualTransferStateStr = await this._redisClient.hGet(key, 'individualItem_' + individualTranferId);
-            if(individualTransferStateStr) {
-                return JSON.parse(individualTransferStateStr) as IndividualTransferState;
-            } else {
-                this._logger.error('Error loading individual trandfer from redis - for key: ' + key);
-                throw (new Error('Error loading individual trandfer from redis'));
-            }
+            return this._data[key][this.individualTransferKeyPrefix + individualTranferId] as IndividualTransferState;
         } catch (err) {
-            this._logger.error(err, 'Error loading individual trandfer from redis - for key: ' + key);
+            this._logger.error(err, 'Error getting individual tranfer from memory - for key: ' + key);
             throw (err);
         }
     }
@@ -163,9 +147,9 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
         }
         const key: string = this.keyWithPrefix(bulkId);
         try {
-            await this._redisClient.hSet(key, 'individualItem_' + individualTranferId, JSON.stringify(value));
+            this._data[key][this.individualTransferKeyPrefix + individualTranferId] = JSON.stringify(value);
         } catch (err) {
-            this._logger.error(err, `Error storing individual trandfer with ID ${individualTranferId} to redis for key: ${key}`);
+            this._logger.error(err, `Error storing individual tranfer with ID ${individualTranferId} to memory for key: ${key}`);
             throw (err);
         }
     }
@@ -176,10 +160,9 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
         }
         const key: string = this.keyWithPrefix(bulkId);
         try {
-            const isExists = await this._redisClient.exists(key);
-            return isExists === 1;
+            return this._data.hasOwnProperty(key);
         } catch (err) {
-            this._logger.error(err, 'Error getting status from redis - for key: ' + key);
+            this._logger.error(err, 'Error getting status from memory - for key: ' + key);
             throw (err);
         }
     }
@@ -189,16 +172,15 @@ export class RedisBulkTransactionStateRepo implements IBulkTransactionEntityRepo
     }
 
     // TODO: Just for development purpose for now, can be removed later
-    // Warning: consider KEYS as a command that should only be used in production environments with extreme care. It may ruin performance when it is executed against large databases. This command is intended for debugging.
     async getAllBulkIds(): Promise<string[]> {
         if(!this.canCall()) {
             throw (new Error('Repository not ready'));
         }
         try {
-            const allKeys =  await this._redisClient.keys(this.keyPrefix + '*');
+            const allKeys =  Object.keys(this._data).filter(key => key.startsWith(this.keyPrefix));
             return allKeys.map(key => key.replace(this.keyPrefix, ''));
         } catch (err) {
-            this._logger.error(err, 'Error getting all bulk transaction ids from redis');
+            this._logger.error(err, 'Error getting all bulk transaction ids from memory');
             throw (err);
         }
     }
