+++ modules/private-shared-lib/src/types/command_event_handler_options.ts
@@ -22,23 +22,11 @@
  --------------
  ******/
 
-import * as ProcessSDKOutboundBulkRequestHandler from './process_sdk_outbound_bulk_request';
-import * as ProcessSDKOutboundBulkPartyInfoRequestHandler from './process_sdk_outbound_bulk_party_info_request';
-import * as ProcessSDKOutboundBulkPartyInfoRequestCompleteHandler from './process_sdk_outbound_bulk_party_info_request_complete';
-import * as ProcessPartyInfoCallbackHandler from './process_party_info_callback';
-import { CommandEventMessage } from '@mojaloop/sdk-scheme-adapter-private-shared-lib';
-import { ICommandEventHandlerOptions } from '@module-types';
-import { ILogger } from '@mojaloop/logging-bc-public-types-lib';
+'use strict';
+import { IDomainEventProducer } from '../types/infra';
+import { IBulkTransactionEntityRepo }  from './bulk_transaction_entity_repo';
 
-export default  {
-    ...ProcessSDKOutboundBulkRequestHandler,
-    ...ProcessSDKOutboundBulkPartyInfoRequestHandler,
-    ...ProcessSDKOutboundBulkPartyInfoRequestCompleteHandler,
-    ...ProcessPartyInfoCallbackHandler,
-} as {
-    [key: string]: (
-        message: CommandEventMessage,
-        options: ICommandEventHandlerOptions,
-        logger: ILogger,
-    ) => Promise<void>
+export type ICommandEventHandlerOptions = {
+    bulkTransactionEntityRepo: IBulkTransactionEntityRepo
+    domainProducer: IDomainEventProducer
 };
