+++ modules/private-shared-lib/src/types/index.ts
@@ -32,3 +32,5 @@
 'use strict';
 
 export * from './infra';
+export * from './bulk_transaction_entity_repo';
+export * from './command_event_handler_options';
